pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ggz-bot"
version = "0.1.0"
description = "A publishable NoneBot2 QQ bot project centered on GTBot."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["nonebot2", "onebot", "qq-bot", "langchain", "gtbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-status>=0.1.3",
    "nonebot2>=2.3.0",
    "numpy>=1.26.0",
    "Pillow>=10.4.0",
    "pydantic>=2.8.0",
    "qdrant-client>=1.10.0",
    "requests>=2.32.0",
    "rich>=13.7.0",
    "SQLAlchemy>=2.0.0",
    "sympy>=1.13.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
cli = [
    "prompt-toolkit>=3.0.47",
]
dev = [
    "mypy>=1.11.0",
    "ruff>=0.6.9",
]
langchain-adapters = [
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-google-vertexai>=2.0.0",
]
providers = [
    "dashscope>=1.23.0",
    "ggz-bot[langchain-adapters]",
]

[tool.setuptools]
py-modules = ["bot"]

[tool.setuptools.packages.find]
include = ["plugins*", "local_plugins*"]
exclude = ["plugins.chatai*"]

[tool.ruff]
target-version = "py310"
line-length = 120
extend-exclude = [
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    "plugins/chatai",
    "plugins/GTBot/test",
    "plugins/GTBot/tests",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]
ignore = [
    "E402",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"plugins/**/test/*.py" = [
    "E402",
]
"plugins/**/test/test_*.py" = [
    "E402",
]

[tool.mypy]
python_version = "3.10"
files = [
    "bot.py",
    "plugins/GTBot",
]
exclude = [
    '^plugins[\\\\/]chatai($|[\\\\/])',
]
namespace_packages = true
follow_imports = "skip"
pretty = true
show_error_codes = true
ignore_missing_imports = true
check_untyped_defs = true
strict_equality = true
warn_return_any = true
disallow_any_generics = true
warn_unused_ignores = false
warn_redundant_casts = false
disable_error_code = [
    "import-untyped",
]

[[tool.mypy.overrides]]
module = [
    "plugins.*.test.*",
    "plugins.*.tests.*",
]
disable_error_code = [
    "annotation-unchecked",
]


[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = ["nonebot_plugin_status"]
plugin_dirs = ["plugins"]
builtin_plugins = ["echo"]
